Add the `send' command.
This commit adds a new command `send'. It will send the marked/selected
tracks to a given playlist (specified as an argument to the command).
Tab expansion is provided. The -c option will create a new playlist.
The -t option will send the currently playing track, instead of the
presently marked/selected track(s).